refactor node/ for common fn
